Sakuraindica1.0_Apps
This application is still in the Beta testing phase. The purpose of this application is to manage the population data of the Sakura Indica neighborhood. Additional information: this application has valid and registered copyright. In the event of a violation, the developer has the right to prosecute or seek compensation for the copyright and plagiarized ideas.